# dbrsdk-python/src/dbrsdk/models/collectionresponse.py
"""Code generated by Speakeasy (https://speakeasy.com). DO NOT EDIT."""

from __future__ import annotations
from dbrsdk.types import BaseModel, Nullable, UNSET_SENTINEL
from pydantic import model_serializer
from typing_extensions import TypedDict


class CollectionResponseTypedDict(TypedDict):
    id: str
    organization_id: str
    name: str
    description: Nullable[str]
    status: str
    estimated_sales_price: Nullable[float]
    estimated_variable_cost: Nullable[float]
    throughput: float
    created_date: str
    updated_date: str


class CollectionResponse(BaseModel):
    id: str

    organization_id: str

    name: str

    description: Nullable[str]

    status: str

    estimated_sales_price: Nullable[float]

    estimated_variable_cost: Nullable[float]

    throughput: float

    created_date: str

    updated_date: str

    @model_serializer(mode="wrap")
    def serialize_model(self, handler):
        optional_fields = []
        nullable_fields = [
            "description",
            "estimated_sales_price",
            "estimated_variable_cost",
        ]
        null_default_fields = []

        serialized = handler(self)

        m = {}

        for n, f in type(self).model_fields.items():
            k = f.alias or n
            val = serialized.get(k)
            serialized.pop(k, None)

            optional_nullable = k in optional_fields and k in nullable_fields
            is_set = (
                self.__pydantic_fields_set__.intersection({n})
                or k in null_default_fields
            )  # pylint: disable=no-member

            if val is not None and val != UNSET_SENTINEL:
                m[k] = val
            elif val != UNSET_SENTINEL and (
                not k in optional_fields or (optional_nullable and is_set)
            ):
                m[k] = val

        return m